+++ behaviors/clubMatches.py
@@ -1,17 +1,14 @@
 import discord
 from discord.ext import commands
 from discord_slash import cog_ext, SlashContext
-import os 
-from dotenv import load_dotenv
+from getGuilds import guilds
 
-load_dotenv()
-GUILD_ID = int(os.getenv('GUILD_ID'))
 
 class clubMatches(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
 
-    @cog_ext.cog_slash(name="match", description='Let me know if we had a "win" or a "loss".', guild_ids = [GUILD_ID])
+    @cog_ext.cog_slash(name="match", description='Let me know if we had a "win" or a "loss".', guild_ids = guilds)
     async def group_say(self, ctx: SlashContext, result: str):
       win_messages = [
         "Way to rep the club!", "Wow! You guys are on a roll!", "Now that’s how you slam!",
